tests: fix for c-span's 403s
