+++ plugins/domains/c-span.org/c-span.org-options.js
@@ -1,0 +1,8 @@
+module.exports = {
+
+    re: require('./c-span.org.js').re,
+
+    getData: function(url, options) {
+        options.exposeStatusCode = true;
+    }
+};
